+++ front-end/src/businessPage/BizOrderStatus.js
@@ -1,12 +1,7 @@
-import styles from './BizOrderStatus.module.css';
+import ComingSoon from '../common/ComingSoon';
 
 function BizOrderStatus() {
-
-    return(
-        <div>
-
-        </div>
-    )
+    return <ComingSoon title="주문 상태 관리" />;
 }
 
 export default BizOrderStatus;