+++ front-end/src/businessPage/BizReview.js
@@ -1,12 +1,7 @@
-import styles from './BizReview.module.css';
+import ComingSoon from '../common/ComingSoon';
 
 function BizReview() {
-
-    return(
-        <div>
-
-        </div>
-    )
+    return <ComingSoon title="리뷰 관리" />;
 }
 
 export default BizReview;