+++ front-end/src/businessPage/BizTempClose.js
@@ -1,12 +1,7 @@
-import styles from './BizTempClose.module.css';
+import ComingSoon from '../common/ComingSoon';
 
 function BizTempClose() {
-
-    return(
-        <div>
-            
-        </div>
-    )
+    return <ComingSoon title="임시 휴무 설정" />;
 }
 
 export default BizTempClose;